Extract file permission utilities
- Move cert/key write helpers into fs_util module.
- Reuse utilities in ACME flow.
- Keep behavior and tests unchanged.

Part of #49
