+++ src/main.rs
@@ -8,6 +8,7 @@ pub mod acme;
 pub mod config;
 pub mod daemon;
 pub mod eab;
+pub mod fs_util;
 pub mod hooks;
 pub mod profile;
